feat: remove custom regex support (#53)
* feat

* feat
